Assignment 1 of Module 1
The first Assingnment was to show a map alongside the given map and the map must show the place of my city in the window and it must be placed parallel to the given map. I also had to make the map interactive in which the user could zoom in and out. I also had to change the background color of the applet and play around wih other functions. It was an easy assignment as i already have worked with applets before.