Assignment 2 of Module 3
This assignment required me to populate the map with live earthquake data taken from a website and populate it on the map based on the Magnitude of each earthquake. So for each class the marker on the map would have a different color and size and then it would be shown on the map. I also had to make an Index of sort, showing which color means what. For example, a Big red marker meant the earthquake had a magnitude of 5.0+, a medium size marker with yellow color had a magnitude of 4.0+ as so on. The main challenge was the understand the different classes dependent on one another to fianlly get the properties of each marker on the map. Surely I'm now understanding the importance of Objects!