Final Assignment
The remaining part of the assignment was to add your own extension for the project. I chose to add the functionality that by clicking at any icon on the Map Key you can filter that type of data and only that is visible. Additionally it also supports multiple clicks which allow users to have more than one data type selected and view that